+++ packages/danger_plugin_dart_test/lib/src/danger_dart_test_reporter.dart
@@ -1,0 +1,5 @@
+import 'package:danger_plugin_dart_test/src/models/danger_dart_error_case.dart';
+
+abstract class DangerDartTestReporter {
+  void reportToDanger(List<DangerDartErrorCase> errorCases);
+}