+++ packages/danger_plugin_dart_test/lib/src/models/danger_dart_error_case.dart
@@ -1,0 +1,8 @@
+class DangerDartErrorCase {
+  final String fileName;
+  final int lineNo;
+  final String testName;
+  final String message;
+
+  DangerDartErrorCase({this.fileName, this.lineNo, this.testName, this.message});
+}