+++ packages/danger_plugin_dart_test/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.5
+
+- Revarmp architecture
+
+## 0.0.4
+
+- Make `limitMessageCharsPerLine` default at null.
+
+## 0.0.3
+
+- Add parameter `limitMessageCharsPerLine` and default at 1000.
+
 ## 0.0.2
 
 - Fix test report case error and golden test error.